TVGUI-4 Reorganising RegisterUser into components folder
